NelPlo {tseries}	R Documentation

Nelson–Plosser Macroeconomic Time Series

Description

These are the extended Nelson-Plosser Data.

Usage

data(NelPlo)

Format

14 macroeconomic time series: cpi, ip, gnp.nom, vel, emp, int.rate, nom.wages, gnp.def, money.stock, gnp.real, stock.prices, gnp.capita, real.wages, and unemp and the joint series NelPlo.

Details

The series are of various lengths but all end in 1988. The data set contains the following series: consumer price index, industrial production, nominal GNP, velocity, employment, interest rate, nominal wages, GNP deflator, money stock, real GNP, stock prices (S&P500), GNP per capita, real wages, unemployment.

Source

Journal of Business and Economic Statistics data archive http://www.amstat.org/publications/jbes/

References

G. Koop and M. F. J. Steel (1994): A Decision-Theoretic Analysis of the Unit-Root Hypothesis using Mixtures of Elliptical Models. Journal of Business and Economic Statistics, 12, 95–107.

[Package tseries version 0.10-45 Index]