garch-methods {tseries}	R Documentation

Methods for Fitted GARCH Models

Description

Methods for fitted GARCH model objects.

Usage

## S3 method for class 'garch'
predict(object, newdata, genuine = FALSE, ...)
## S3 method for class 'garch'
coef(object, ...)
## S3 method for class 'garch'
vcov(object, ...)
## S3 method for class 'garch'
residuals(object, ...)
## S3 method for class 'garch'
fitted(object, ...)
## S3 method for class 'garch'
print(x, digits = max(3, getOption("digits") - 3), ...)
## S3 method for class 'garch'
plot(x, ask = interactive(), ...)
## S3 method for class 'garch'
logLik(object, ...)

Arguments

object, x
an object of class "garch"; usually, a result of a call to garch.

newdata
a numeric vector or time series to compute GARCH predictions. Defaults to eval(parse(text=object$series)).

genuine
a logical indicating whether a genuine prediction should be made, i.e., a prediction for which there is no target observation available.

digits
see printCoefmat.

ask
Should the plot method work interactively? See interactive.

...
further arguments passed to or from other methods.

Details

predict returns +/- the conditional standard deviation predictions from a fitted GARCH model.

coef returns the coefficient estimates.

vcov the associated covariance matrix estimate (outer product of gradients estimator).

residuals returns the GARCH residuals, i.e., the time series used to fit the model divided by the computed conditional standard deviation predictions for this series. Under the assumption of conditional normality the residual series should be i.i.d. standard normal.

fitted returns +/- the conditional standard deviation predictions for the series which has been used to fit the model.

plot graphically investigates normality and remaining ARCH effects for the residuals.

logLik returns the log-likelihood value of the GARCH(p, q) model represented by object evaluated at the estimated coefficients. It is assumed that first max(p, q) values are fixed.

Value

For predict a bivariate time series (two-column matrix) of predictions.

For coef, a numeric vector, for residuals and fitted a univariate (vector) and a bivariate time series (two-column matrix), respectively.

For plot and print, the fitted GARCH model object.

Author(s)

A. Trapletti

[Package tseries version 0.10-43 Index]