ice.river {tseries}	R Documentation

Icelandic River Data

Description

Contains the Icelandic river data as presented in Tong (1990), pages 432–440.

Usage

data(ice.river)

Format

4 univariate time series flow.vat, flow.jok, prec, and temp, each with 1095 observations and the joint series ice.river.

Details

The series are daily observations from Jan. 1, 1972 to Dec. 31, 1974 on 4 variables: flow.vat, mean daily flow of Vatnsdalsa river (cms), flow.jok, mean daily flow of Jokulsa Eystri river (cms), prec, daily precipitation in Hveravellir (mm), and mean daily temperature in Hveravellir (deg C).

These datasets were introduced into the literature in a paper by Tong, Thanoon, and Gudmundsson (1985).

Source

Time Series Data Library: https://robjhyndman.com/TSDL/

References

H. Tong (1990): Non-Linear Time Series, A Dynamical System Approach. Oxford University Press, Oxford.

H. Tong, B. Thanoon, and G. Gudmundsson (1985): Threshold time series modelling of two Icelandic riverflow systems. Water Resources Bulletin, 21, 651–661.

[Package tseries version 0.10-45 Index]