jarque.bera.test {tseries}	R Documentation

Jarque–Bera Test

Description

Tests the null of normality for x using the Jarque-Bera test statistic.

Usage

jarque.bera.test(x)

Arguments

x
a numeric vector or time series.

Details

This test is a joint statistic using skewness and kurtosis coefficients.

Missing values are not allowed.

Value

A list with class "htest" containing the following components:

statistic
the value of the test statistic.

parameter
the degrees of freedom.

p.value
the p-value of the test.

method
a character string indicating what type of test was performed.

data.name
a character string giving the name of the data.

Author(s)

A. Trapletti

References

J. B. Cromwell, W. C. Labys and M. Terraza (1994): Univariate Tests for Time Series Models, Sage, Thousand Oaks, CA, pages 20–22.

Examples

x <- rnorm(100)  # null
jarque.bera.test(x)

x <- runif(100)  # alternative
jarque.bera.test(x)

[Package tseries version 0.10-34 Index]