na.remove {tseries}	R Documentation

NA Handling Routines for Time Series

Description

Observations with missing values in some of the variables are removed.

Usage

na.remove(object, ...)

Arguments

object
a numeric matrix, vector, univariate, or multivariate time series.

...
further arguments to be passed to or from methods.

Details

For na.remove.ts this changes the “intrinsic” time scale. It is assumed that both, the new and the old time scale are synchronized at the first and the last valid observation. In between, the new series is equally spaced in the new time scale.

Value

An object without missing values. The attribute "na.removed" contains the indices of the removed missing values in object.

Author(s)

A. Trapletti

See Also

na.omit, na.fail

Examples

x<-ts(c(5453.08,5409.24,5315.57,5270.53, # one and a half week stock index
        5211.66,NA,NA,5160.80,5172.37))  # data including a weekend
na.remove(x)  # eliminate weekend and introduce ``business'' time scale

[Package tseries version 0.10-43 Index]