po.test {tseries}	R Documentation

Phillips–Ouliaris Cointegration Test

Description

Computes the Phillips-Ouliaris test for the null hypothesis that x is not cointegrated.

Usage

po.test(x, demean = TRUE, lshort = TRUE) 

Arguments

x
a matrix or multivariate time series.

demean
a logical indicating whether an intercept is included in the cointegration regression or not.

lshort
a logical indicating whether the short or long version of the truncation lag parameter is used.

Details

The Phillips-Perron Z(alpha) statistic for a unit root in the residuals of the cointegration regression is computed, see also pp.test. The unit root is estimated from a regression of the first variable (column) of x on the remaining variables of x without a constant and a linear trend. To estimate sigma^2 the Newey-West estimator is used. If lshort is TRUE, then the truncation lag parameter is set to trunc(n/100), otherwise trunc(n/30) is used. The p-values are interpolated from Table Ia and Ib, p. 189 of Phillips and Ouliaris (1990). If the computed statistic is outside the table of critical values, then a warning message is generated.

The dimension of x is restricted to six variables. Missing values are not handled.

Value

A list with class "htest" containing the following components:

statistic
the value of the test statistic.

parameter
the truncation lag parameter.

p.value
the p-value of the test.

method
a character string indicating what type of test was performed.

data.name
a character string giving the name of the data.

Author(s)

A. Trapletti

References

P. C. B. Phillips and S. Ouliaris (1990): Asymptotic Properties of Residual Based Tests for Cointegration. Econometrica 58, 165–193.

See Also

pp.test

Examples

x <- ts(diffinv(matrix(rnorm(2000),1000,2)))  # no cointegration
po.test(x)

x <- diffinv(rnorm(1000))
y <- 2.0-3.0*x+rnorm(x,sd=5)
z <- ts(cbind(x,y))  # cointegrated
po.test(z)

[Package tseries version 0.10-43 Index]