read.matrix {tseries}	R Documentation

Read Matrix Data

Description

Reads a matrix data file.

Usage

read.matrix(file, header = FALSE, sep = "", skip = 0)

Arguments

file
the name of the file which the data are to be read from.

header
a logical value indicating whether the file contains the names of the columns as its first line.

sep
the field separator character. Values on each line of the file are separated by this character.

skip
the number of lines of the data file to skip before beginning to read data.

Details

Usually each row of the file represents an observation and each column contains a variable. The first row possibly contains the names of the variables (columns).

read.matrix might be more efficient than read.table for very large data sets.

Author(s)

A. Trapletti

See Also

read.table.

Examples

x <- matrix(0, 10, 10)
write(x, "test", ncolumns=10)
x <- read.matrix("test")
x
unlink("test")

[Package tseries version 0.10-43 Index]