seqplot.ts {tseries}	R Documentation

Plot Two Time Series

Description

Plot two time series on the same plot frame.

Usage

seqplot.ts(x, y, colx = "black", coly = "red", typex = "l",
           typey = "l", pchx = 1, pchy = 1, ltyx = "solid",
           ltyy = "solid", oma = c(6, 0, 5, 0), ann = par("ann"),
           xlab = "Time", ylab = deparse(substitute(x)), main = NULL)

Arguments

x, y
the time series.

colx, coly
color code or name for the x and y series, see colors, palette.

typex, typey
what type of plot should be drawn for the x and y series, see plot.

pchx, pchy
character or integer code for kind of points/lines for the x and y series.

ltyx, ltyy
line type code for the x and y series, see lines.

oma
a vector giving the size of the outer margins in lines of text, see par.

ann
annotate the plots? See par.

xlab, ylab
titles for the x and y axis.

main
an overall title for the plot.

Details

Unlike plot.ts the series can have different time bases, but they should have the same frequency. Unlike ts.plot the series can be plotted in different styles and for multivariate x and y the common variables are plotted together in a separate array element.

Value

None.

Author(s)

A. Trapletti

See Also

ts, plot.ts

Examples

data(USeconomic)
x <- ts.union(log(M1), log(GNP), rs, rl)
m.ar <- ar(x, method = "ols", order.max = 5)
y <- predict(m.ar, x, n.ahead = 200, se.fit = FALSE)
seqplot.ts(x, y)

[Package tseries version 0.10-45 Index]