sterling {tseries}	R Documentation

Sterling Ratio

Description

This function computes the Sterling ratio of the univariate time series (or vector) x.

Usage

sterling(x)

Arguments

x
a numeric vector or univariate time series corresponding to a portfolio's cumulated returns.

Details

The Sterling ratio is defined as a portfolio's overall return divided by the portfolio's maxdrawdown statistic. In finance the Sterling Ratio represents a measure of the portfolio's risk-adjusted return.

Value

a double representing the Sterling ratio.

Author(s)

A. Trapletti

See Also

maxdrawdown, sharpe

Examples

data(EuStockMarkets)
dax <- log(EuStockMarkets[,"DAX"])
ftse <- log(EuStockMarkets[,"FTSE"])
sterling(dax)
sterling(ftse)

[Package tseries version 0.10-45 Index]