summary.arma {tseries}	R Documentation

Summarizing ARMA Model Fits

Description

Methods for creating and printing summaries of ARMA model fits.

Usage

## S3 method for class 'arma'
summary(object, ...)
## S3 method for class 'summary.arma'
print(x, digits = max(3, getOption("digits") - 3),
      signif.stars = getOption("show.signif.stars"), ...)

Arguments

object
an object of class "arma"; usually, a result of a call to arma.

x
an object of class "summary.arma"; usually, a result of a call to the summary method for objects of class "arma".

digits, signif.stars
see printCoefmat.

...
further arguments passed to or from other methods.

Details

The summary method computes the asymptotic standard errors of the coefficient estimates from the numerically differentiated Hessian matrix approximation. The AIC is computed from the conditional sum-of-squared errors and not from the true maximum likelihood function. That may be problematic.

Value

A list of class "summary.arma".

See Also

arma

[Package tseries version 0.10-43 Index]