summary.garch {tseries}	R Documentation

Summarizing GARCH Model Fits

Description

Methods for creating and printing summaries of GARCH model fits.

Usage

## S3 method for class 'garch'
summary(object, ...)
## S3 method for class 'summary.garch'
print(x, digits = max(3, getOption("digits") - 3),
      signif.stars = getOption("show.signif.stars"), ...)

Arguments

object
an object of class "garch"; usually, a result of a call to garch.

x
an object of class "summary.garch"; usually, a result of a call to the summary method for objects of class "garch".

digits, signif.stars
see printCoefmat.

...
further arguments passed to or from other methods.

Details

summary computes the asymptotic standard errors of the coefficient estimates from an outer-product approximation of the Hessian evaluated at the estimates, see Bollerslev (1986). It furthermore tests the residuals for normality and remaining ARCH effects, see jarque.bera.test and Box.test.

Value

A list of class "summary.garch".

References

T. Bollerslev (1986): Generalized Autoregressive Conditional Heteroscedasticity. Journal of Econometrics 31, 307–327.

See Also

garch

[Package tseries version 0.10-45 Index]