tcm {tseries}	R Documentation

Monthly Yields on Treasury Securities

Description

This data set contains monthly 1 year, 3 year, 5 year, and 10 year yields on treasury securities at constant, fixed maturity.

Usage

data(tcm)

Format

4 univariate time series tcm1y, tcm3y, tcm5y, and tcm10y and the joint series tcm.

Details

The yields at constant fixed maturity have been constructed by the Treasury Department, based on the most actively traded marketable treasury securities.

Source

U.S. Fed http://www.federalreserve.gov/Releases/H15/data.htm

[Package tseries version 0.10-45 Index]