bev {tseries}	R Documentation

Beveridge Wheat Price Index, 1500–1869.

Description

Contains the well-known Beveridge Wheat Price Index which gives annual price data from 1500 to 1869, averaged over many locations in western and central Europe.

Usage

data(bev)

Format

A univariate time series with 370 observations. The object is of class "ts".

Details

This data provides an example of long memory time series which has the appearance of being nonstationary in levels and yet also appears overdifferenced. See, e.g., Baillie (1996).

Source

Time Series Data Library: https://robjhyndman.com/TSDL/

References

R. T. Baillie (1996): Long Memory Processes and Fractional Integration in Econometrics. Journal of Econometrics, 73, 5–59.

[Package tseries version 0.10-45 Index]