camp {tseries}	R Documentation

Mount Campito Yearly Treering Data, -3435–1969.

Description

Contains annual tree-ring measurements from Mount Campito from 3426 BC through 1969 AD.

Usage

data(camp)

Format

A univariate time series with 5405 observations. The object is of class "ts".

Details

This series is a standard example for the concept of long memory time series.

The data was produced and assembled at the Tree Ring Laboratory at the University of Arizona, Tuscon.

Source

Time Series Data Library: https://robjhyndman.com/TSDL/

[Package tseries version 0.10-45 Index]