tcmd {tseries}	R Documentation

Daily Yields on Treasury Securities

Description

This data set contains daily 1 year, 3 year, 5 year, and 10 year yields on treasury securities at constant, fixed maturity.

Usage

data(tcmd)

Format

4 univariate time series tcm1yd, tcm3yd, tcm5yd, and tcm10yd and the joint series tcmd.

Details

The yields at constant fixed maturity have been constructed by the Treasury Department, based on the most actively traded marketable treasury securities.

Daily refers to business days, i.e., weekends and holidays are eliminated.

Source

U.S. Fed http://www.federalreserve.gov/Releases/H15/data.htm

[Package tseries version 0.10-43 Index]